Now with Balanced Parentheses!
it's not regex, its not posix, but it can fly… finally balanced
parentheses to fix all the () issue. it's ugly but it's fast and works
with all kinds of combinations.

it basically just counts open vs. closed and if they are even again it
knows exactly where the condition is and does not cut into html or cut
off too much.